Split dashboard bundle so the main page stops shipping salon data
The build was a single 870 kB chunk: Layout imported fmtDate from
lib/data, which pulls in the 388 kB salon.json, and every page (with
Recharts) was eagerly imported. Locale/format helpers move to
lib/format.ts (no data imports), lib/data.ts re-exports them so salon
pages are unchanged, and App.tsx loads all routes via lazy() behind a
Suspense fallback.

The entry chunk is now 220 kB (71 kB gzip); the salon dataset (227 kB)
and Recharts (337 kB) load only on the routes that need them. nginx
serves index.html with no-cache so a stale document can't request
chunk filenames that a redeploy removed. salon-dashboard/README.md
replaces the Vite boilerplate and documents the split, including the
rule that always-loaded modules must import lib/format, not lib/data.
