+++ deploy/nginx-vest-smr.conf
@@ -24,6 +24,12 @@ server {
         add_header Cache-Control "no-cache";
     }
 
+    # index.html не кешируем: имена чанков меняются при каждом деплое,
+    # иначе браузер будет запрашивать уже удалённые /assets/*.js
+    location = /index.html {
+        add_header Cache-Control "no-cache";
+    }
+
     location /assets/ {
         expires 30d;
         add_header Cache-Control "public, immutable";
