+++ salon-dashboard/README.md
@@ -1,32 +1,60 @@
-# React + TypeScript + Vite
+# salon-dashboard
 
-This template provides a minimal setup to get React working in Vite with HMR and some Oxlint rules.
+Фронтенд vest-smr.ru: дашборд мониторинга ИИ-индустрии (корень сайта) и аналитика салона (раздел `/salon`).
+React 19 + TypeScript + Vite, стили — Tailwind, графики — Recharts, роутинг — wouter.
 
-Currently, two official plugins are available:
+## Команды
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react) uses [Oxc](https://oxc.rs)
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react-swc) uses [SWC](https://swc.rs/)
-
-## React Compiler
-
-The React Compiler is not enabled on this template because of its impact on dev & build performances. To add it, see [this documentation](https://react.dev/learn/react-compiler/installation).
-
-## Expanding the Oxlint configuration
-
-If you are developing a production application, we recommend enabling type-aware lint rules by installing `oxlint-tsgolint` and editing `.oxlintrc.json`:
-
-```json
-{
-  "$schema": "./node_modules/oxlint/configuration_schema.json",
-  "plugins": ["react", "typescript", "oxc"],
-  "options": {
-    "typeAware": true
-  },
-  "rules": {
-    "react/rules-of-hooks": "error",
-    "react/only-export-components": ["warn", { "allowConstantExport": true }]
-  }
-}
+```bash
+npm ci          # установка зависимостей
+npm run dev     # dev-сервер с HMR
+npm run build   # tsc -b + vite build → dist/
+npm run lint    # oxlint
+npm run preview # локальный просмотр собранного dist/
 ```
 
-See the [Oxlint rules documentation](https://oxc.rs/docs/guide/usage/linter/rules) for the full list of rules and categories.
+## Структура
+
+```
+src/
+  App.tsx              маршруты (все страницы — lazy)
+  components/Layout.tsx шапка, навигация, футер
+  components/ui/       card, badge, button
+  data/report.ts       данные отчёта по ИИ: метрики, модели, цены, рекомендации
+  data/salon.json      выгрузка salonbackup (1127 визитов), ~388 КБ
+  hooks/useNews.ts     подгрузка /data/news.json с фолбэком на вшитый дайджест
+  lib/format.ts        формат чисел и дат, без зависимостей от данных
+  lib/data.ts          датасет салона: entries, expenses, masters, прайс
+  pages/               страницы разделов ИИ и салона
+```
+
+## Разбиение бандла
+
+Страницы подключены через `lazy()`, поэтому Recharts и `salon.json` не попадают
+в стартовую загрузку главной страницы: `/` тянет ~228 КБ (~75 КБ gzip), выгрузка
+салона (~227 КБ) грузится только при переходе в `/salon`.
+
+Чтобы это не сломалось:
+
+- **не импортируйте `lib/data` из `Layout` или других модулей, которые грузятся всегда** —
+  через него в основной чанк попадёт `salon.json`. Для формата чисел и дат
+  используйте `lib/format` (`lib/data` реэкспортит его для страниц салона);
+- новые страницы добавляйте в `App.tsx` тоже через `lazy(() => import(...))`.
+
+После сборки проверяйте, что в выводе `npm run build` нет предупреждения
+о чанках больше 500 КБ, кроме вендорного чанка Recharts.
+
+## Данные новостей
+
+`useNews` в рантайме запрашивает `/data/news.json`; если файла нет — показывает
+дайджест, вшитый в `data/report.ts`. Файл публикуется на сервер отдельно
+воркфлоу `.github/workflows/update-news.yml` (см. `deploy/README.md`), поэтому
+обновление новостей не требует пересборки фронтенда. В nginx `/data/news.json`
+и `index.html` отдаются с `Cache-Control: no-cache`, а `/assets/` — с
+долгим кешем, так как имена файлов содержат хеш.
+
+## Деплой
+
+Сборка выкладывается в `/var/www/vest-smr` на VM воркфлоу
+`.github/workflows/deploy-vest-smr.yml`, конфиг nginx — `deploy/nginx-vest-smr.conf`.
+Вариант в контейнере: `Dockerfile` + `nginx.container.conf` (порт 8080).
