+++ salon-dashboard/nginx.container.conf
@@ -14,6 +14,12 @@ server {
         add_header Cache-Control "no-cache";
     }
 
+    # index.html не кешируем: имена чанков меняются при каждом деплое,
+    # иначе браузер будет запрашивать уже удалённые /assets/*.js
+    location = /index.html {
+        add_header Cache-Control "no-cache";
+    }
+
     location /assets/ {
         expires 30d;
         add_header Cache-Control "public, immutable";
